Break out style datatypes into separate crate
