+++ azul-css-parser/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "azul-css"
+name = "azul-css-parser"
 version = "0.1.0"
 authors = ["Felix Schütt <felix.schuett@maps4print.com>"]
 description = '''
@@ -16,7 +16,8 @@ exclude = ["assets/*", "doc/*", "examples/*"]
 autoexamples = false
 
 [dependencies]
-azul                    = { path = "../azul",   default-features = false }
-simplecss               = { version = "0.1.0",  default-features = false }
-euclid                  = { version = "0.19",   default-features = false }
-app_units               = { version = "0.7",    default-features = false } # TODO: Remove once webrender PR is merged
+#azul                    = { path = "../azul",   default-features = false }
+azul-style              = { path = "../azul-style", default-features = false }
+simplecss               = { version = "0.1.0",      default-features = false }
+#euclid                  = { version = "0.19",   default-features = false }
+#app_units               = { version = "0.7",    default-features = false } # TODO: Remove once webrender PR is merged
