+++ azul-css-parser/src/dom.rs
@@ -1,4 +1,4 @@
-use azul::prelude::NodeTypePath;
+use azul_style::NodeTypePath;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
 pub enum NodeTypePathParseError<'a> {
@@ -11,7 +11,7 @@ impl_display!{ NodeTypePathParseError<'a>, {
 
 /// Parses the node type from a CSS string such as `"div"` => `NodeTypePath::Div`
 pub fn node_type_path_from_str(data: &str) -> Result<NodeTypePath, NodeTypePathParseError> {
-    use azul::prelude::NodeTypePath::*;
+    use azul_style::NodeTypePath::*;
     match data {
         "div" => Ok(Div),
         "p" => Ok(P),
