+++ azul-style/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "azul-style"
+version = "0.1.0"
+authors = ["Felix Schütt <felix.schuett@maps4print.com>"]
+description = '''
+    Common datatypes used for styling applications
+    using the Azul desktop GUI framework
+'''
+documentation = "https://docs.rs/azul"
+homepage = "https://azul.rs/"
+keywords = ["gui", "GUI", "user interface", "svg", "graphics", "style" ]
+categories = ["gui"]
+license = "MIT"
+repository = "https://github.com/maps4print/azul"
+readme = "README.md"
+exclude = ["assets/*", "doc/*", "examples/*"]
+autoexamples = false
+
+[dependencies]
+euclid                  = { version = "0.19",   default-features = false }
+
+# Windows: Use nativefiledialog because of better Unicode support for file browsers
+[target.'cfg(not(target_os = "linux"))'.dependencies]
+nfd = { version = "0.0.4", default-features = false }
+webrender = { git = "https://github.com/servo/webrender", rev = "92235d1fc4ff4c4fb4e1ed709eb3136e96c786b5", default-features = false }
+
+# Linux: Use webrender with freetype feature enabled
+[target.'cfg(target_os = "linux")'.dependencies]
+webrender = { git = "https://github.com/servo/webrender", rev = "92235d1fc4ff4c4fb4e1ed709eb3136e96c786b5" }
