+++ azul-style/src/macros.rs
@@ -1,0 +1,31 @@
+/// Implement the `From` trait for any type.
+/// Example usage:
+/// ```
+/// enum MyError<'a> {
+///     Bar(BarError<'a>)
+///     Foo(FooError<'a>)
+/// }
+/// 
+/// impl_from!(BarError<'a>, Error::Bar);
+/// impl_from!(BarError<'a>, Error::Bar);
+/// 
+/// ```
+macro_rules! impl_from {
+    // From a type with a lifetime to a type which also has a lifetime
+    ($a:ident<$c:lifetime>, $b:ident::$enum_type:ident) => {
+        impl<$c> From<$a<$c>> for $b<$c> {
+            fn from(e: $a<$c>) -> Self {
+                $b::$enum_type(e)
+            }
+        }
+    };
+
+    // From a type without a lifetime to a type which also does not have a lifetime
+    ($a:ident, $b:ident::$enum_type:ident) => {
+        impl From<$a> for $b {
+            fn from(e: $a) -> Self {
+                $b::$enum_type(e)
+            }
+        }
+    };
+}
