+++ azul/src/app_state.rs
@@ -9,6 +9,7 @@ use image::ImageError;
 #[cfg(feature = "image_loading")]
 use images::ImageType;
 use rusttype::Font;
+use azul_style::{FontId, StyleFontSize, PixelValue, StyleLetterSpacing};
 use {
     FastHashMap,
     text_cache::TextId,
@@ -18,7 +19,6 @@ use {
     traits::Layout,
     app_resources::AppResources,
     font::FontError,
-    style_properties::{FontId, StyleFontSize, PixelValue, StyleLetterSpacing},
     error::ClipboardError,
     daemon::{Daemon, DaemonId, TerminateDaemon},
 };
