+++ azul/src/text_layout.rs
@@ -3,13 +3,13 @@
 use webrender::api::LayoutPixel;
 use euclid::{TypedRect, TypedSize2D, TypedPoint2D};
 use rusttype::{Font, Scale, GlyphId};
+use azul_style::{
+    StyleTextAlignmentHorz, StyleFontSize, StyleBackgroundColor, StyleLetterSpacing,
+    FontId, StyleTextAlignmentVert, StyleLineHeight, LayoutOverflow
+};
 use {
     app_resources::AppResources,
     text_cache::TextInfo,
-    style_properties::{
-        StyleTextAlignmentHorz, StyleFontSize, StyleBackgroundColor, StyleLetterSpacing,
-        FontId, StyleTextAlignmentVert, StyleLineHeight, LayoutOverflow
-    },
     text_cache::{TextId, TextCache},
 };
 
@@ -802,7 +802,7 @@ fn words_to_left_aligned_glyphs<'a>(
                 }
 
                 for glyph in &word.glyphs {
-                    use style_properties::PT_TO_PX;
+                    use azul_style::PT_TO_PX;
 
                     // vertical_advance is in px
                     let mut new_glyph = *glyph;
@@ -891,7 +891,7 @@ fn align_text_horz(
     glyphs: &mut [GlyphInstance],
     line_breaks: &[(usize, f32)]
 ) {
-    use style_properties::StyleTextAlignmentHorz::*;
+    use azul_style::StyleTextAlignmentHorz::*;
 
     // Text alignment is theoretically very simple:
     //
